adverbs moved to own folder.
Due to my previous attempt going up in smoke I'm doing like one of these at a time. :D

The ` target_include_directories` is working. All these files will look in `jsrc` as if they were in that folder.
So you to include `a.h` from `j.h` you need `#include “adverbs/a.h”`.
But to go the other way you don’t need to change it. `j.h` from `a.h` is just `` `#include “j.h”`

You can have multiple directories in `target_include_directories` but it seems to break when you have directories in the same tree like `./jsrc/adverbs` and `./jsrc`. But something like `./jsrc/adverbs` and `./jsrc/misc` should work.
