+++ jsrc/adverbs/a.c
