+++ jsrc/adverbs/ab.c
