+++ jsrc/adverbs/ai.c
