+++ jsrc/adverbs/am.c
