+++ jsrc/adverbs/am1.c
