+++ jsrc/adverbs/amn.c
