+++ jsrc/adverbs/ao.c
