+++ jsrc/adverbs/ap.c
