+++ jsrc/adverbs/ar.c
