+++ jsrc/adverbs/as.c
