+++ jsrc/adverbs/au.c
