+++ jsrc/j.h
@@ -822,7 +822,7 @@ if(likely(z<3)){_zzt+=z; z=(I)&oneone; _zzt=_i&3?_zzt:(I*)z; z=_i&2?(I)_zzt:z; z
 #include "vq.h" 
 #include "vx.h" 
 #include "vz.h"
-#include "a.h"
+#include "adverbs/a.h"
 #include "s.h"
